Convert types

+++ orchestrator/cosmos_gravity/src/send.rs
@@ -136,7 +136,7 @@ async fn __send_messages(
 
     // compute the fee as fee=ceil(gas_limit * gas_price)
 
-    let fee_amount: f64 = args.fee.gas_limit * gas_price.0;
+    let fee_amount: f64 = args.fee.gas_limit as f64 * gas_price.0;
     let fee_amount: u64 = fee_amount.abs().ceil() as u64;
     let fee_amount = Coin {
         denom: gas_price.1,
@@ -185,7 +185,7 @@ pub async fn send_messages(
     args.fee.gas_limit = cmp::max(gas_limit as u64, 500000 * messages.len() as u64);
 
     // compute the fee as fee=ceil(gas_limit * gas_price)
-    let fee_amount: f64 = args.fee.gas_limit * gas_price.0;
+    let fee_amount: f64 = args.fee.gas_limit as f64 * gas_price.0;
     let fee_amount: u64 = fee_amount.abs().ceil() as u64;
     let fee_amount = Coin {
         denom: gas_price.1,
